refactor material imports; region factory
